+++ lib/elixir_script/module_systems.ex
@@ -4,16 +4,12 @@ defmodule ElixirScript.ModuleSystems do
     ElixirScript.ModuleSystems.ES6
   end
 
-  def import_module(module_names, from, env) when is_list(module_names) do
-    module_system.import_module(module_names, from, env)
+  def import_namespace_module(module_name, from, env) do
+    module_system.import_namespace_module(module_name, from, env)
   end
 
   def import_module(module_name, from, env) do
     module_system.import_module(module_name, from, env)
-  end
-
-  def import_module(module_name, %ElixirScript.Translator.LexicalScope{} = env) do
-    module_system.import_module(module_name, env)
   end
 
   def import_module(import_name, from) do